chore: Move table func trait
Moves it close to where it's implemented. The 'builtins' file is already a bit big.
