+++ crates/sqlbuiltins/src/functions/table/iceberg.rs
@@ -6,7 +6,6 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use super::table_location_and_opts;
-use crate::builtins::TableFunc;
 use async_trait::async_trait;
 pub(crate) use data_files::*;
 use datafusion::arrow::array::{Int32Builder, Int64Builder, StringBuilder, UInt64Builder};
@@ -20,6 +19,8 @@ use datasources::lake::storage_options_into_object_store;
 use protogen::metastore::types::catalog::{FunctionType, RuntimePreference};
 pub(crate) use scan::*;
 pub(crate) use snapshots::*;
+
+use super::TableFunc;
 
 fn box_err<E>(err: E) -> ExtensionError
 where
