+++ crates/sqlbuiltins/src/functions/table/snowflake.rs
@@ -10,7 +10,8 @@ use datafusion_ext::functions::{FuncParamValue, TableFuncContextProvider};
 use datasources::snowflake::{SnowflakeAccessor, SnowflakeDbConnection, SnowflakeTableAccess};
 use protogen::metastore::types::catalog::{FunctionType, RuntimePreference};
 
-use crate::builtins::{ConstBuiltinFunction, TableFunc};
+use super::TableFunc;
+use crate::functions::ConstBuiltinFunction;
 
 #[derive(Debug, Clone, Copy)]
 pub struct ReadSnowflake;
